Group store-related types
