+++ src/Language/Mimsa/Backend/Javascript.hs
@@ -26,7 +26,7 @@ import Language.Mimsa.Backend.NormaliseConstructors
 import Language.Mimsa.Printer
 import Language.Mimsa.Types.AST
 import Language.Mimsa.Types.Identifiers
-import Language.Mimsa.Types.ResolvedTypeDeps
+import Language.Mimsa.Types.Store.ResolvedTypeDeps
 
 ----
 newtype Javascript = Javascript Text
