+++ src/Language/Mimsa/Project/Usages.hs
@@ -6,13 +6,9 @@ import Data.Set (Set)
 import qualified Data.Set as S
 import Language.Mimsa.Project.Persistence (getCurrentBindings)
 import Language.Mimsa.Store.ResolvedDeps (resolveDeps)
-import Language.Mimsa.Types.Bindings
-import Language.Mimsa.Types.ExprHash
 import Language.Mimsa.Types.Identifiers
 import Language.Mimsa.Types.Project
-import Language.Mimsa.Types.ResolvedDeps
 import Language.Mimsa.Types.Store
-import Language.Mimsa.Types.StoreExpression
 import Language.Mimsa.Types.Usage
 
 findUsages :: Project ann -> ExprHash -> Either UsageError (Set Usage)
