+++ src/Language/Mimsa/Project/Versions.hs
@@ -16,13 +16,11 @@ import Language.Mimsa.Actions
 import Language.Mimsa.Project.Usages
 import Language.Mimsa.Types.AST
 import Language.Mimsa.Types.Error
-import Language.Mimsa.Types.ExprHash
 import Language.Mimsa.Types.Identifiers
 import Language.Mimsa.Types.MonoType
 import Language.Mimsa.Types.Project
 import Language.Mimsa.Types.ResolvedExpression
 import Language.Mimsa.Types.Store
-import Language.Mimsa.Types.StoreExpression
 import Language.Mimsa.Types.Usage
 import Language.Mimsa.Types.VersionedMap
