+++ src/Language/Mimsa/Types.hs
@@ -2,22 +2,17 @@ module Language.Mimsa.Types
   ( module Language.Mimsa.Types.TypeConstructor,
     module Language.Mimsa.Types.AST,
     module Language.Mimsa.Types.Identifiers,
-    module Language.Mimsa.Types.ExprHash,
     module Language.Mimsa.Types.Usage,
     module Language.Mimsa.Types.Store,
     module Language.Mimsa.Types.MonoType,
     module Language.Mimsa.Types.Scheme,
     module Language.Mimsa.Types.Substitutions,
-    module Language.Mimsa.Types.ResolvedDeps,
     module Language.Mimsa.Types.Environment,
     module Language.Mimsa.Types.Error,
     module Language.Mimsa.Types.Scope,
     module Language.Mimsa.Types.Swaps,
     module Language.Mimsa.Types.UniVar,
-    module Language.Mimsa.Types.Bindings,
-    module Language.Mimsa.Types.StoreExpression,
     module Language.Mimsa.Types.Project,
-    module Language.Mimsa.Types.TypeBindings,
     module Language.Mimsa.Types.VersionedMap,
     module Language.Mimsa.Types.ResolvedExpression,
     module Language.Mimsa.Types.SubstitutedExpression,
@@ -25,23 +20,18 @@ module Language.Mimsa.Types
 where
 
 import Language.Mimsa.Types.AST
-import Language.Mimsa.Types.Bindings
 import Language.Mimsa.Types.Environment
 import Language.Mimsa.Types.Error
-import Language.Mimsa.Types.ExprHash
 import Language.Mimsa.Types.Identifiers
 import Language.Mimsa.Types.MonoType
 import Language.Mimsa.Types.Project
-import Language.Mimsa.Types.ResolvedDeps
 import Language.Mimsa.Types.ResolvedExpression
 import Language.Mimsa.Types.Scheme
 import Language.Mimsa.Types.Scope
 import Language.Mimsa.Types.Store
-import Language.Mimsa.Types.StoreExpression
 import Language.Mimsa.Types.SubstitutedExpression
 import Language.Mimsa.Types.Substitutions
 import Language.Mimsa.Types.Swaps
-import Language.Mimsa.Types.TypeBindings
 import Language.Mimsa.Types.TypeConstructor
 import Language.Mimsa.Types.UniVar
 import Language.Mimsa.Types.Usage
