+++ src/Language/Mimsa/Types/Error/ResolverError.hs
@@ -5,9 +5,9 @@ module Language.Mimsa.Types.Error.ResolverError where
 import qualified Data.Map as M
 import qualified Data.Text as T
 import Language.Mimsa.Printer (Printer (prettyPrint))
-import Language.Mimsa.Types.Bindings (Bindings (Bindings))
 import Language.Mimsa.Types.Identifiers (Name, TyCon)
-import Language.Mimsa.Types.TypeBindings
+import Language.Mimsa.Types.Store.Bindings (Bindings (Bindings))
+import Language.Mimsa.Types.Store.TypeBindings
   ( TypeBindings (TypeBindings),
   )
