+++ src/Language/Mimsa/Types/ResolvedExpression.hs
@@ -4,7 +4,7 @@ import Language.Mimsa.Types.AST
 import Language.Mimsa.Types.Identifiers
 import Language.Mimsa.Types.MonoType
 import Language.Mimsa.Types.Scope
-import Language.Mimsa.Types.StoreExpression
+import Language.Mimsa.Types.Store.StoreExpression
 import Language.Mimsa.Types.Swaps
 
 data ResolvedExpression ann
