+++ src/Language/Mimsa/Types/Store/Bindings.hs
@@ -2,15 +2,15 @@
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
 {-# LANGUAGE OverloadedStrings #-}
 
-module Language.Mimsa.Types.Bindings where
+module Language.Mimsa.Types.Store.Bindings where
 
 import qualified Data.Aeson as JSON
 import Data.Map (Map)
 import qualified Data.Map as M
 import qualified Data.Text as T
 import Language.Mimsa.Printer
-import Language.Mimsa.Types.ExprHash
 import Language.Mimsa.Types.Identifiers
+import Language.Mimsa.Types.Store.ExprHash
 
 -- a list of names to hashes
 newtype Bindings = Bindings {getBindings :: Map Name ExprHash}
