+++ src/Language/Mimsa/Types/Store/ResolvedDeps.hs
@@ -1,14 +1,14 @@
 {-# LANGUAGE OverloadedStrings #-}
 
-module Language.Mimsa.Types.ResolvedDeps where
+module Language.Mimsa.Types.Store.ResolvedDeps where
 
 import Data.Map (Map)
 import qualified Data.Map as M
 import Data.Text.Prettyprint.Doc
 import Language.Mimsa.Printer
-import Language.Mimsa.Types.ExprHash
 import Language.Mimsa.Types.Identifiers
-import Language.Mimsa.Types.StoreExpression
+import Language.Mimsa.Types.Store.ExprHash
+import Language.Mimsa.Types.Store.StoreExpression
 
 newtype ResolvedDeps a
   = ResolvedDeps
