+++ src/Language/Mimsa/Types/Store/ResolvedTypeDeps.hs
@@ -1,12 +1,12 @@
-module Language.Mimsa.Types.ResolvedTypeDeps where
+module Language.Mimsa.Types.Store.ResolvedTypeDeps where
 
 import Data.Map (Map)
 import qualified Data.Map as M
 import Data.Text.Prettyprint.Doc
 import Language.Mimsa.Printer
 import Language.Mimsa.Types.AST
-import Language.Mimsa.Types.ExprHash
 import Language.Mimsa.Types.Identifiers
+import Language.Mimsa.Types.Store.ExprHash
 
 newtype ResolvedTypeDeps
   = ResolvedTypeDeps
