+++ src/Language/Mimsa/Types/Store/StoreExpression.hs
@@ -2,14 +2,14 @@
 {-# LANGUAGE DeriveFunctor #-}
 {-# LANGUAGE DeriveGeneric #-}
 
-module Language.Mimsa.Types.StoreExpression where
+module Language.Mimsa.Types.Store.StoreExpression where
 
 import qualified Data.Aeson as JSON
 import GHC.Generics
 import Language.Mimsa.Types.AST
-import Language.Mimsa.Types.Bindings
 import Language.Mimsa.Types.Identifiers
-import Language.Mimsa.Types.TypeBindings
+import Language.Mimsa.Types.Store.Bindings
+import Language.Mimsa.Types.Store.TypeBindings
 
 -- a storeExpression contains the AST Expr
 -- and a map of names to hashes with further functions inside
