+++ src/Language/Mimsa/Types/Store/TypeBindings.hs
@@ -2,15 +2,15 @@
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
 {-# LANGUAGE OverloadedStrings #-}
 
-module Language.Mimsa.Types.TypeBindings where
+module Language.Mimsa.Types.Store.TypeBindings where
 
 import qualified Data.Aeson as JSON
 import Data.Map (Map)
 import qualified Data.Map as M
 import qualified Data.Text as T
 import Language.Mimsa.Printer (Printer (prettyPrint))
-import Language.Mimsa.Types.ExprHash (ExprHash)
 import Language.Mimsa.Types.Identifiers (TyCon)
+import Language.Mimsa.Types.Store.ExprHash (ExprHash)
 
 -- a list of names to hashes
 newtype TypeBindings = TypeBindings {getTypeBindings :: Map TyCon ExprHash}
