+++ src/Language/Mimsa/Types/Usage.hs
@@ -4,8 +4,8 @@ module Language.Mimsa.Types.Usage where
 
 import qualified Data.Text as T
 import Language.Mimsa.Printer
-import Language.Mimsa.Types.ExprHash
 import Language.Mimsa.Types.Identifiers
+import Language.Mimsa.Types.Store.ExprHash
 
 data Usage
   = Transient Name ExprHash
